Improved backend

+++ backend/models/expense_split.py
@@ -4,7 +4,7 @@ class ExpenseSplitModel(db.Model):
     __tablename__ = "expense_splits"
 
     id = db.Column(db.Integer, primary_key=True)
-    amount = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Numeric(10, 2), nullable=False)
 
     expense_id = db.Column(db.Integer, db.ForeignKey("expenses.id"), nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
